+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Preserve per-folder grouping, sorting and view type ("Use for this folder only") when exporting and importing settings ([#133])
 
 ## [1.13.1] - 2026-02-14
 ### Changed
@@ -266,6 +268,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#29]: https://github.com/FossifyOrg/Gallery/issues/29
 [#62]: https://github.com/FossifyOrg/Gallery/issues/62
 [#128]: https://github.com/FossifyOrg/Gallery/issues/128
+[#133]: https://github.com/FossifyOrg/Gallery/issues/133
 [#166]: https://github.com/FossifyOrg/Gallery/issues/166
 [#199]: https://github.com/FossifyOrg/Gallery/issues/199
 [#237]: https://github.com/FossifyOrg/Gallery/issues/237